pyproject.toml
[project]
name = "risklab-diagnostics-agent"
version = "0.1.0"
description = "Compact decision-support system for model-risk and stress-testing questions over a frozen three-document corpus."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.33.0",
    "langchain-core>=1.2.20",
    "langgraph>=1.1.3",
    "langsmith>=0.7.22",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.2.2",
    "pymupdf4llm>=1.27.2.2",
    "pyyaml>=6.0.3",
    "rank-bm25>=0.2.2",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "sentence-transformers>=5.3.0",
    "typer>=0.24.1",
]

[project.scripts]
risklab = "app.cli:main"

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv]
package = true

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.7",
]
